Separation witness related code
